+++ nym-vpn-core/crates/nym-vpn-lib-types/src/lib.rs
@@ -44,6 +44,7 @@ mod privy;
 mod rpc_requests;
 mod service;
 mod socks5;
+mod split_tunnel;
 mod tunnel_event;
 mod tunnel_state;
 #[cfg(feature = "uniffi-bindings")]
@@ -97,6 +98,7 @@ pub use service::{
     TargetState, VpnServiceConfig, VpnServiceInfo,
 };
 pub use socks5::{EnableSocks5Request, HttpRpcSettings, Socks5Settings, Socks5State, Socks5Status};
+pub use split_tunnel::{SplitTunnelExcludedProcess, SplitTunnelExcludedProcessList};
 pub use tunnel_event::{
     BandwidthEvent, ConnectionEvent, ConnectionStatisticsEvent, MixnetEvent, SphinxPacketRates,
     TunnelEvent,